+++ src/js/utils/constants.js
@@ -1,0 +1,2 @@
+
+export const IS_PRODUCTION = process.env.NODE_ENV === 'production';