+++ src/js/utils/debug.js
@@ -1,15 +1,5 @@
 /* eslint-disable no-console */
-/**
- * Console utils
- * @module utils/console
- */
-
-/**
- * Decorate a message
- * @param {string} msg - The message to log.
- * @param {string} decorator - Console message decorator.
- */
-const decorateMsg = (msg, decorator = '📐') => `${decorator} ${msg}`;
+import { IS_PRODUCTION } from './constants';
 
 /**
  * Log a decorated message to the console
@@ -17,10 +7,19 @@ const decorateMsg = (msg, decorator = '📐') => `${decorator} ${msg}`;
  * @param {string} decorator - Console message decorator.
  */
 const message = (method, ...messages) => {
-  if (!!console && !!console[method]) {
-    const [first, ...rest] = messages;
-    console[method](decorateMsg(first), ...rest);
+  if (IS_PRODUCTION) { return; }
+
+  if (console && console[method]) {
+    console[method](...messages);
   }
+};
+
+/**
+ * Clear the console
+ */
+export const clear = () => {
+  if (IS_PRODUCTION) { return; }
+  console.clear();
 };
 
 /**
@@ -28,44 +27,25 @@ const message = (method, ...messages) => {
  * @param {string} msg - The message to log.
  * @param {string} decorator - Console message decorator.
  */
-const error = (...messages) => message('error', ...messages);
+export const error = (...messages) => message('error', ...messages);
 
 /**
  * Info wrapper
  * @param {string} msg - The message to log.
  * @param {string} decorator - Console message decorator.
  */
-const info = (...messages) => message('info', ...messages);
+export const info = (...messages) => message('info', ...messages);
 
 /**
  * Log wrapper
  * @param {string} msg - The message to log.
  * @param {string} decorator - Console message decorator.
  */
-const log = (...messages) => message('log', ...messages);
+export const log = (...messages) => message('log', ...messages);
 
 /**
  * Warn wrapper
  * @param {string} msg - The message to log.
  * @param {string} decorator - Console message decorator.
  */
-const warn = (...messages) => message('warn', ...messages);
-
-export const debug = (mode, ...messages) => {
-  if (process.env.NODE_ENV === 'production') { return; }
-
-  switch (mode) {
-    case 'error':
-      error(...messages);
-      break;
-    case 'info':
-      info(...messages);
-      break;
-    case 'warn':
-      warn(...messages);
-      break;
-    default:
-      log(...messages);
-      break;
-  }
-};
+export const warn = (...messages) => message('warn', ...messages);