+++ src/js/utils/index.js
@@ -1,8 +1,0 @@
-/**
- * Utility Functions
- * @module utils
- */
-
-export { debug } from './debug';
-export { ready } from './ready';
-export { htmlToElement, htmlToElements } from './template';